# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Lumen API",
        "description": "API for the Lumen project",
        "version": "1.0.0"
    },
    "paths": {
...
        "/materias": {
            "get": {
                "tags": [
                    "Materias"
                ],
                "summary": "Listar todas las materias",
                "operationId": "3e87d7d7e4512cf8d0c2208a27ab32cd",
                "responses": {
                    "200": {
                        "description": "Listado de materias",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Materia"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al obtener las materias"
                    }
                }
            },
            "post": {
                "tags": [
                    "Materias"
                ],
                "summary": "Crear una nueva materia",
                "operationId": "9eb7e9f92071de08e0b9c7bdcd49b6cd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "maxLength": 100,
                                        "example": "Cálculo Diferencial"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Materia creada exitosamente"
                    },
                    "400": {
                        "description": "Error de validación"
                    },
                    "500": {
                        "description": "Error al crear la materia"
                    }
                }
            }
        },
        "/materias/{id}": {
            "get": {
                "tags": [
                    "Materias"
                ],
                "summary": "Obtener una materia por su ID",
                "operationId": "7d890d7d9d28b94eb6034c430caf4109",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la materia",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Materia encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Materia"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Materia no encontrada"
                    },
                    "500": {
                        "description": "Error al obtener la materia"
                    }
                }
            },
            "put": {
                "tags": [
                    "Materias"
                ],
                "summary": "Actualizar una materia existente",
                "operationId": "1248df134213e2bdcd3571407b6581af",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la materia",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "maxLength": 100,
                                        "example": "Cálculo Integral"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Materia actualizada exitosamente"
                    },
                    "400": {
                        "description": "Error de validación"
                    },
                    "500": {
                        "description": "Error al actualizar la materia"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Materias"
                ],
                "summary": "Eliminar una materia",
                "operationId": "8a8cc6baa4259fcaf4d3ba8258f5c14d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la materia",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Materia eliminada exitosamente"
                    },
                    "400": {
                        "description": "Error al eliminar la materia (e.g., dependencias)"
                    },
                    "500": {
                        "description": "Error al eliminar la materia"
                    }
                }
            }
        },
        "/materias/asignar-docente": {
            "post": {
                "tags": [
                    "Materias"
                ],
                "summary": "Asignar un docente a una materia",
                "operationId": "97fdd38cdbc82829a5447dc8d9cf052c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_materia",
                                    "id_docente"
                                ],
                                "properties": {
                                    "id_materia": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "id_docente": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Docente asignado a la materia exitosamente"
                    },
                    "403": {
                        "description": "No tienes permiso para asignar docentes a esta materia"
                    },
                    "400": {
                        "description": "Error de validación"
                    },
                    "500": {
                        "description": "Error al asignar el docente a la materia"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Materia": {
                "title": "Materia",
                "description": "Modelo de Materia",
                "properties": {
                    "id_materia": {
                        "description": "ID de la materia",
                        "type": "integer",
                        "readOnly": "true"
                    },
                    "nombre": {
                        "description": "Nombre de la materia",
                        "type": "string",
                        "maxLength": 255
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}